requirement.txt
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
contextlib2==21.6.0
future==0.18.2
google-api-core==2.7.3
google-api-python-client==2.47.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.56.0
httplib2==0.20.4
icloudpd==1.7.2
idna==3.3
importlib-metadata==4.11.3
keyring==23.5.0
keyrings.alt==4.1.0
oauthlib==3.2.0
piexif==1.1.3
progress==1.6
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyicloud==1.0.0
pyicloud-ipd==0.10.1
pyparsing==3.0.8
python-dateutil==2.8.1
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
schema==0.7.2
six==1.16.0
tqdm==4.56.0
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.9
zipp==3.8.0
